Refactor FigureSupplier to use Color enum instead of String for color
