+++ src/main/java/core/basesyntax/Main.java
@@ -1,5 +1,7 @@
 package core.basesyntax;
 
+import core.basesyntax.logic.Figure;
+
 public class Main {
     public static void main(String[] args) {
