+++ src/main/java/core/basesyntax/figure/Square.java
@@ -1,10 +1,13 @@
-package core.basesyntax;
+package core.basesyntax.figure;
+
+import core.basesyntax.Color;
+import core.basesyntax.logic.Figure;
 
 public class Square extends Figure {
 
     private final double side;
 
-    public Square(String color, double side) {
+    public Square(Color color, double side) {
         super(color);
         this.side = side;
     }
