+++ src/main/java/core/basesyntax/logic/AreaCalculator.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package core.basesyntax.logic;
 
 public interface AreaCalculator {
     double getArea();
