+++ src/main/java/core/basesyntax/logic/PrintableFigure.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package core.basesyntax.logic;
 
 public interface PrintableFigure {
     void draw();
